Refactor: Member 패키지 삭제 (Member -> Session 이관)
