--- session-service/src/main/java/com/sparta/moim/session/member/application/dto/command/GetMemberCommand.java
@@ -1,6 +1,0 @@
-package com.sparta.moim.session.member.application.dto.command;
-
-import java.util.UUID;
-
-public record GetMemberCommand(UUID sessionId) {
-}
