--- session-service/src/main/java/com/sparta/moim/session/member/application/event/feign/SessionInternalService.java
@@ -1,9 +1,0 @@
-package com.sparta.moim.session.member.application.event.feign;
-
-import java.util.UUID;
-
-public interface SessionInternalService {
-  void getSessionValidate(UUID sessionId);
-  void getSessionValidateTime(UUID sessionId);
-  void getSessionValidateOpenStatus(UUID sessionId);
-}
