--- session-service/src/main/java/com/sparta/moim/session/member/domain/enums/MemberType.java
@@ -1,5 +1,0 @@
-package com.sparta.moim.session.member.domain.enums;
-
-public enum MemberType {
-  PUBLISHER, GENERAL
-}
