--- session-service/src/main/java/com/sparta/moim/session/member/presentation/dto/response/GetMemberListResponse.java
@@ -1,13 +1,0 @@
-package com.sparta.moim.session.member.presentation.dto.response;
-
-import com.sparta.moim.session.member.application.dto.result.GetMemberListResult;
-import java.util.UUID;
-
-public record GetMemberListResponse(
-    UUID id,
-    String type
-) {
-  public GetMemberListResponse(GetMemberListResult member) {
-    this(member.id(), member.type());
-  }
-}
