--- session-service/src/main/java/com/sparta/moim/session/session/application/dto/DeleteSessionCommand.java
@@ -1,6 +1,0 @@
-package com.sparta.moim.session.session.application.dto;
-
-import java.util.UUID;
-
-public record DeleteSessionCommand(UUID sessionId, String username) {
-}
