+++ session-service/src/main/java/com/sparta/moim/session/session/application/dto/command/DeleteSessionCommand.java
@@ -1,0 +1,6 @@
+package com.sparta.moim.session.session.application.dto.command;
+
+import java.util.UUID;
+
+public record DeleteSessionCommand(UUID sessionId, String username) {
+}
