+++ session-service/src/main/java/com/sparta/moim/session/session/application/dto/command/RemoveMemberCommand.java
@@ -1,0 +1,7 @@
+package com.sparta.moim.session.session.application.dto.command;
+
+import java.util.List;
+import java.util.UUID;
+
+public record RemoveMemberCommand(UUID sessionId, List<UUID> members) {
+}
