+++ session-service/src/main/java/com/sparta/moim/session/session/domain/enums/MemberType.java
@@ -1,0 +1,5 @@
+package com.sparta.moim.session.session.domain.enums;
+
+public enum MemberType {
+  PUBLISHER, GENERAL
+}
