+++ session-service/src/main/java/com/sparta/moim/session/session/presentation/dto/request/RemoveMemberRequest.java
@@ -1,0 +1,14 @@
+package com.sparta.moim.session.session.presentation.dto.request;
+
+import com.sparta.moim.session.session.application.dto.command.RemoveMemberCommand;
+import jakarta.validation.constraints.NotNull;
+import java.util.List;
+import java.util.UUID;
+
+public record RemoveMemberRequest(
+    @NotNull List<UUID> members
+) {
+  public RemoveMemberCommand toCommand(UUID sessionId) {
+    return new RemoveMemberCommand(sessionId, members);
+  }
+}
