+++ session-service/src/main/java/com/sparta/moim/session/session/presentation/dto/response/GetMemberListResponse.java
@@ -1,0 +1,13 @@
+package com.sparta.moim.session.session.presentation.dto.response;
+
+import com.sparta.moim.session.session.application.dto.result.GetMemberListResult;
+import java.util.UUID;
+
+public record GetMemberListResponse(
+    UUID id,
+    String type
+) {
+  public GetMemberListResponse(GetMemberListResult member) {
+    this(member.id(), member.type());
+  }
+}
