+++ session-service/src/main/java/com/sparta/moim/session/shared/configuration/RedisLeaveStreamConfig.java
@@ -1,7 +1,7 @@
 package com.sparta.moim.session.shared.configuration;
 
 
-import com.sparta.moim.session.member.infrastructure.event.listener.redis.RedisStreamLeaveListener;
+import com.sparta.moim.session.session.infrastructure.event.listener.redis.RedisStreamLeaveListener;
 import java.time.Duration;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
