+++ session-service/src/test/java/com/sparta/moim/session/member/presentation/controller/internal/InternalMemberControllerTest.java
@@ -14,8 +14,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.moim.session.member.application.MemberService;
-import com.sparta.moim.session.member.application.dto.result.GetMemberListResult;
+import com.sparta.moim.session.session.application.service.MemberService;
+import com.sparta.moim.session.session.application.dto.result.GetMemberListResult;
+import com.sparta.moim.session.session.presentation.controller.internal.InternalMemberController;
 import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
